BATCHEE-156 only log a warning on CNFE
There is one use case which might be ok:
If someone works with different implementations, e.g. different JPA providers
then the Exceptions might be different depending on implementations.
In that case the batch developer might want to add Exceptions from all the supported
implementations. And of course only one of them is available at runtime in that case.
The other enlisted classes will likely throw a NCFE.